Pagination component splitted in two. UI changes.
