+++ frontend/app/pages/archive/index.vue
@@ -7,7 +7,10 @@
     <ArchiveHeader/>
     <ArchiveSearchSection/>
     <ArchiveBody/>
-    <ArchivePagination/>
+    <div class="max-w-5xl mx-auto px-4 pt-0 pb-6 flex items-center justify-between">
+      <ArchivePagination />
+      <ArchivePageJumper />
+    </div>
   </div>
 </template>
